pyproject.toml
[project]
name = "palm-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "streamlit>=1.48.0",
]
